requirements.txt
oracledb>=1.4.2
python-dotenv>=1.0,<2.0

PyYAML>=6.0,<7.0
-e ../python_data_platform